user module can do basic database crud operations

+++ entities/base.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Base struct {
-	ID        uint           `json:"id" gorm:"primarykey"`
+	ID        uint64         `json:"id" gorm:"primarykey"`
 	CreatedAt time.Time      `json:"createdAt"`
 	UpdatedAt time.Time      `json:"updatedAt"`
 	DeletedAt gorm.DeletedAt `json:"deletedAt" gorm:"index"`
